+++ modules/Grades/StudentGrades.php
@@ -99,7 +99,7 @@ if ( UserStudentID()
 
 				$gradebook_config[$staff_id]['WEIGHT'] = issetVal( $gradebook_config[$staff_id]['WEIGHT'] );
 
-				// @since 8.7.1 Exclude 0 points assignments from Ungraded count.
+				// @since 8.8 Exclude 0 points assignments from Ungraded count.
 				$sql = "SELECT s.STUDENT_ID,gt.ASSIGNMENT_TYPE_ID,
 				sum(" . db_case( array( 'gg.POINTS', "'-1'", "'0'", 'gg.POINTS' ) ) . ") AS PARTIAL_POINTS,
 				sum(" . db_case( array( 'gg.POINTS', "'-1'", "'0'", 'ga.POINTS' ) ) . ") AS PARTIAL_TOTAL,